Game is working. Needs refactoring and adjusting unit tests after refactoring Board to Game and GameStats
